move `Republisher` to a separate file
Extract the `Republisher` structure and related logic from `system.go` (leaving
that file mainly with the `Root` logic, and hence now renamed as `root.go`) to a
new `repub.go` file. Even though the MFS root has support for republishing
updates the logic behind those two structures is mostly decoupled, that should
be reflected in the source code organization.
